backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
pinecone-client==3.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy>=1.26.0
